+++ Editor/Drawers/SaintsRowDrawer/SaintsRowAttributeDrawerUIToolkit.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Reflection;
 using SaintsField.Editor.Drawers.DateTimeDrawer;
 using SaintsField.Editor.Drawers.EnumFlagsDrawers.EnumToggleButtonsDrawer;
+using SaintsField.Editor.Drawers.SaintsInterfacePropertyDrawer;
 using SaintsField.Editor.Drawers.TimeSpanDrawer;
 using SaintsField.Editor.Drawers.TreeDropdownDrawer;
 using SaintsField.Editor.Playa;
@@ -177,7 +178,8 @@ namespace SaintsField.Editor.Drawers.SaintsRowDrawer
                 {
                     label = label[..^"__Saints Serialized__".Length];
                 }
-                VisualElement renderSerializedActual = RenderSerializedActual(saintsSerializedActual, label, property, info, saintsSerializedActual.ElementType, inHorizontalLayout, SerializedUtils.GetFieldInfoAndDirectParent(property).parent);
+                // Debug.Log($"{info.Name}/{property.propertyPath}/{saintsSerializedActual.Name}/{saintsSerializedActual.ElementType}");
+                VisualElement renderSerializedActual = RenderSerializedActual(saintsSerializedActual, label, property, (FieldInfo)info, saintsSerializedActual.ElementType, inHorizontalLayout, SerializedUtils.GetFieldInfoAndDirectParent(property).parent);
                 root.Add(renderSerializedActual);
                 return;
             }
@@ -266,7 +268,7 @@ namespace SaintsField.Editor.Drawers.SaintsRowDrawer
         }
 
         private static VisualElement RenderSerializedActual(SaintsSerializedActualAttribute saintsSerializedActual,
-            string label, SerializedProperty property, MemberInfo serInfo, Type targetType, bool inHorizontalLayout, object parent)
+            string label, SerializedProperty property, FieldInfo serInfo, Type targetType, bool inHorizontalLayout, object parent)
         {
             // Debug.Log(property.propertyPath);
             Attribute[] attributes = ReflectCache.GetCustomAttributes(serInfo);
@@ -314,7 +316,12 @@ namespace SaintsField.Editor.Drawers.SaintsRowDrawer
                     return TreeDropdownAttributeDrawer.RenderSerializedActual((ISaintsAttribute)flagsTreeDropdownAttribute ?? flagsDropdownAttribute, label, property, targetType);
                     // return null;
                 }
+                case SaintsPropertyType.Interface:
+                {
+                    return SaintsInterfaceDrawer.RenderSerializedActual(label, property, targetType, attributes, inHorizontalLayout, serInfo, parent);
+                }
                 case SaintsPropertyType.Undefined:
+                case SaintsPropertyType.ClassOrStruct:
                 default:
                     return null;
             }